requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

SQLAlchemy==2.0.32
pydantic==2.8.2
pydantic-settings==2.4.0

google-auth==2.34.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.143.0

APScheduler==3.10.4
python-dotenv==1.0.1

openai==1.40.6

jinja2==3.1.4
gunicorn
psycopg2-binary